avalanche/requirements.txt
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.8.22
geopandas==0.12.2
google-api-core==2.11.0
google-auth==2.16.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
lxml==4.9.2
munch==2.5.0
numpy==1.24.1
packaging==23.0
pandas==1.5.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
rsa==4.9
shapely==2.0.0
six==1.16.0
urllib3==1.26.14
